+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - [add feature-env plugin & add spring cloud gateway staining plugin](https://github.com/Tencent/spring-cloud-tencent/pull/532)
 - [feature:add @ConditionalOnConfigReflectEnabled annotation](https://github.com/Tencent/spring-cloud-tencent/pull/538)
 - [fix:set error handler named EnhancedRestTemplateReporter for RestTemplate](https://github.com/Tencent/spring-cloud-tencent/pull/544)
+- [Optimize: add switch for report call result and default false](https://github.com/Tencent/spring-cloud-tencent/pull/549)